requirements.txt
Flask==3.1.3
python-dotenv==1.2.1
requests==2.32.5
flasgger==0.9.7.1
SQLAlchemy==2.0.47
PyMySQL==1.1.2
click==8.3.1
flask-jwt-extended==4.7.1
bcrypt==4.3.0
numpy
pandas
scikit-learn
cryptography
